+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.4.2
+
+- Added `path.clear` global config option to optionally clear the global scratch directory during pipeline cleanup (use with caution).
+- Rewrote unit tests in `tests/`
+
 ## 2.4.1
 
 - Added options for `nexus md analyze`, giving customizable cpptraj outputs